feat: implement `proofset list` cli/client/api
allows users to list the proofsets their node is currently configured
with. This will come in handy w/ later PRs where we init a node and a
proofset automatically if one doesn't exist when a "full" node starts.
The intention here is to use this API on startup in a later PR.