fix: replica transfer conclude invocation (#241)
This fixes the invocation made to the upload service to communicate the
completion of the `blob/replica/transfer` task.

We use `client.Execute(...)` to make a `ucan/conclude` invocation
instead of hand rolling and sending an agent message.

This also adds code to inspect the receipt for a failure. Since the UCAN
service is RPC style, it will still return a HTTP 200 even if the
receipt contains an error message. So where previously this would have
succeeded if the upload service sent an failure receipt, it will now
fail.

fixes https://github.com/storacha/piri/issues/234