feat: UCAN authorized retrievals (#238)
This PR adds UCAN authorized retrievals via `space/content/retrieve`
invocations, as specified
[here](https://github.com/storacha/specs/blob/main/w3-retrieval.md).

It adds a go-ucanto retrieval server, and with a handler for
`space/content/retrieve`. The server will serve the blob from the
blobstore provided it was allocated in the space referenced by the
invocation.

In PDP mode, blobs are stored in a different store (or in a completely
different process), keyed by piece hash so we use an adapter to allow
the blobs keyed by piece hash to be accessed via user hash.

BREAKING CHANGE: allocations are now keyed by blob digest and space
instead of blob digest and invocation CID.