feat: implement egress tracking (#245, #246, #258)
Ref: #174 

Implement an `EgressBatchStore` that allows appending
`space/content/retrieve` receipts to a batch. When the batch is above
the max, it is flushed automatically.

Things that are missing and will be implemented in follow up PRs:
- sending the batch to the egress tracking service in a
`space/egress/track` invocation
- exposing an `http.FileSystem` and adding the corresponding
`http.FileServer` endpoint

Once these are done, I'll hook it up in Alan's `space/content/retrieve`
handler.

---------

Co-authored-by: ash <alan@storacha.network>