feat: check consolidated records and clean receipt batches (#264)
Ref: #174 

Poll the etracker service periodically to fetch
`space/egress/consolidate` receipts. This requires storing the
consolidate invocation link sent in the `space/egress/track` receipt as
an effect. It will be used later to fetch the corresponding receipt from
the etracker receipts endpoint.

The original track invocation is also stored so that it can be used when
checking the consolidate receipt (validation of the receipt still TODO).

Both the track invocation and the consolidate invocation link are stored
in a new, datastore-based `ConsolidationStore`, keyed by the batch CID
under two different namespaces.

Once consolidation has happened and the receipt has been confirmed to be
correct, the batch can be removed from the journal.

---------

Co-authored-by: ash <alan@storacha.network>