Upgrade to PDP Verifier v2.1.0 (#256)
This PR updates the contract bindings to meet the interface of the most
recently deployed PDPVerifier v2.1.0 smart contracts. The diff is
misleading, the 5,000+ LOC are from code gen from the contract ABI.

Biggest changes here, aside from a new contract, is that we now use
PieceCIDv2 cids for contract interaction as it requires. Left a bunch of
TODOs and REVIEW comments here for discussion.

The second biggest change is the removal of the "service" contract from
the code, currently we use only the PDP Verifier contract, and use hard
coded values of challenge window and proving period. Data Sets are
created in the Verifier without a listener/record-keeper, the contract
doesn't require a listener, and we use hard coded values for proving
period and challenge window.