fix: `ucan/conclude` invocation facts (#360)
### Problem

The upload service executes a `ucan/conclude` invocation sent from Piri.

The receipt for that invocation is issued by the server, but receipts
are not intelligent to the fact that they are for a `ucan/conclude`
invocation. They simply call `iterateIPLDBlocks()` (JS version of
`Export()`) on their `ran` invocation, which exports the invocation, and
any blocks linked to from _caveats and facts_. This will export the
receipt block (the block linked from `receipt` in `nb`), but _not_ any
blocks linked to from effects in that receipt.

The consequence is that the agent message for the `ucan/conclude`
invocation includes these blocks, but the agent message for the receipt
does not. However, both messages are indexed as containing a receipt for
the invocation (because they do), but only one of them includes the
supporting blocks.

The client needs access to those supporting blocks, but sometimes they
are not returned when the client asks for a receipt for the
`blob/replica/transfer` task 😱 Since CIDs are basically random bytes,
the message that is selected to be returned is random, so sometimes
you'll get the message with the supporting blocks and other times you
won't.

### Solution

This PR fixes the `ucan/conclude` invocation made by Piri to link to all
the receipt blocks from facts, so that when the invocation is executed
the issued receipt will export the concluded receipt block _as well as_
blocks the receipt links to (because they are linked from the invocation
facts).

This is not unusual. We already do this for all other conclude
invocations see:

https://github.com/storacha/upload-service/blob/a36b3289a77e529ea29f22bc6152be00520ccb34/packages/upload-api/src/ucan/conclude.js#L76-L102
and

https://github.com/storacha/upload-service/blob/a36b3289a77e529ea29f22bc6152be00520ccb34/packages/upload-client/src/blob/add.js#L134-L160

Below is an example agent message - the receipt for a `ucan/conclude`
invocation. I have decoded all the blocks included in the message, but
as you can see there are CIDs to other blocks that are not included.
Conversely, when you look at the agent message for the invocation, it
_does_ include these blocks.

<img width="1215" height="305" alt="Screenshot 2025-11-24 at 13 22 58"
src="https://github.com/user-attachments/assets/4df54b58-d1b4-4b42-9aec-7851985c6acb"
/>