fix: do not display usage when init error (#363)
I *think* this stops cobra from printing the usage text when an error
occurs. I think this makes sense here, espeicially when we get a few
steps in and it is clear the usage of the command is correct.

Helpfully undocumented, but you can click through from here to some
removed docs for it:
https://github.com/spf13/cobra/issues/225

---------

Co-authored-by: frrist <forrest@storacha.network>