docs: update version numbers and add prod network setup guide (#372)
Updates the version numbers in the docs, adds the `--network` flag to
`piri init` docs and adds a guide for getting setup on the production
network, either as a new node or mirgating from Staging.

I have purposely made the production network setup guide a separate
document and left the existing guide focused solely on bootstrapping to
staging.